alternative appoarch to keep json the same
